feat: update scrutinizer
