+++ .scrutinizer.yml
@@ -8,7 +8,7 @@ build:
   nodes:
     analysis:
       environment:
-        php: 7.4
+        php: 8.2
         postgresql: false
         redis: false
         mongodb: false
