+++ src/Broker/Broker.php
@@ -82,12 +82,7 @@ class Broker
         $this->broker = $broker;
         $this->secret = $secret;
 
-        // check if PHP version >=8.0
-        if (phpversion() >= '8.0.0') {
-            $this->state = new Cookies8();
-        } else {
-            $this->state = new Cookies();
-        }
+        $this->state = new Cookies();
     }
 
     /**
